backend/requirements.txt
ace_tools==0.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
attrs==25.3.0
autograder-py==0.6.12
canvasapi==3.3.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
discord==2.3.2
discord.py==2.5.2
distro==1.9.0
fastapi==0.115.12
flake8==7.1.2
fonttools==4.53.1
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
griffe==1.7.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
idna==3.7
jiter==0.5.0
kiwisolver==1.4.5
matplotlib==3.9.2
mccabe==0.7.0
mcp==1.6.0
multidict==6.4.3
numpy==2.0.1
openai==1.76.0
openai-agents==0.0.13
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.3.6
propcache==0.3.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pyflakes==3.2.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.3.3
starlette==0.46.2
tqdm==4.66.5
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uv==0.6.17
uvicorn==0.34.2
yarl==1.20.0
